Charm dispersal 🔕: Move Death's Door to Necromancy only
Death's door is very much not an "always on" charm, it's very powerful
but with major downsides. It works as a continued effect in the
Necromancy school.

If single-school level 8 proves too inviting for this effect it can go
to level 9; invulnerability by courting death is after all a very cool
and thematic necromancy capstone.