Charm dispersal 🔕: Move Song of Slaying to Tmut and adjust
Song of Slaying's effect is nice: you make noise but you punch harder,
the noise draws more dudes, you get a nice spiral, but press your luck
too long and the effect ends.

This commit moves the spell to Transmutations and reflavours it
"Wereblood": transforms some of the caster's blood into that of a
were-beast, giving them a slay bonus on kills. The noise is reduced to
only trigger on a kill, instead of continuously throughout the duration
(still loud casting noise as well).