Charm dispersal 🔕: Spell removals
This commit removes the remaining spells in the charms school: Infusion,
Shroud of Golubria, Spectral Weapon, and Ring of Flames, as well as
Darkness.

They are all removed for similar reasons: in general if one was using
them it was a "cast once then start fighting" or "always keep this up
it's for the best" situation. Neither option fits well with how crawls
MP system works tactically.

Some of these effects (Spectral Weapon, Ring of Flames, and Darkness)
are cool and we don't want to lose them, so subsequent commits will
reintroduce them as items. For this reason, their supporting code is
left intact by this commit.

Infusion offered a "trade MP for a bit of bonus melee damage"; this
could maybe be restored as an amulet, but its hard to see if such an
amulet would be competitive against the other options in the slot. For
now, it gets the axe.

Shroud of Golubria offered "trade 2 MP for effectively +10 HP"; we
already have an item that uses MP to increase effective HP: guardian
spirit! So shroud gets the axe. The name is freed up and could make a
cool unrand armour name with some kind of translocations flavour.