WIP Charm dispersal 🔕: Move Battlesphere to Conjurations only
Battlesphere is in a bit of an awkward place with the positional magic
changes. Unlike other charms the MP cost of battlesphere is
relevant, as its use depends on subsequent casts and it could end up in
a less than useful position.

WIP: Battlesphere needs to be aware of more magic (zap at a target in
range when conjurations are cast). Add that as a squash commit later in
the branch.