Charming item 🔔: Spectral ego
Re-introduce spectral weapons as a weapon ego. The ego is flavoured as
drawing out the ancient tree spirit from the wood the weapon is made out
of, and the strength of the spectral weapon scales with evocations
skill.

The brand spawns on staves, clubs, giant clubs, and giant spiked clubs
as these are "woody enough" flavour wise. Mechanically, staves are
2handers with good base damage, the spectral ego offers a tradeoff of
brand for having a spectral weapon; a similar statement can be made
about giant (spiked) clubs, with the additional remark that in both
cases the shield slot is restricted. Clubs getting the rare spectral
brand is partly flavour and partly fun, even at 27 evo a spectral club
doesn't hit that hard (7 base dam). Randarts of any weapon type can roll
the brand, but it is limited to two-handers. Brand scrolls never give
the brand, which should keep it relatively rare.

Monsters wielding a spectral ego weapon get a spectral weapon (thanks
gammafunk for not deleting the supporting code when Asterion lost the
spell). Spectral weapon doesn't work great on monsters because of
monster AI, but the limited weapon family should keep it to monsters
that are better at catching the player out in the open (Spriggans and
staves).