squash! Charming item 🔔: Amulet of Flames
Charming item 🔔: Salamander Hide Armour

The Salamander hide has been on the "unrands to revise" list for a
while. Ring of Flames is too flashy for a mundane item, this commit
moves the effect to the Salamander hide, giving it rF++, rC--, a fire
enhancer, and a ring of clouds and losing +Rage.